streamlit/requirements.txt
streamlit==1.28.0
kafka-python==2.0.2
pandas==2.1.0
minio==7.1.17
python-dotenv==1.0.0
plotly==5.17.0
pyarrow==13.0.0
trino==0.323.0
sqlalchemy==2.0.0
streamlit-autorefresh==0.0.1
numpy==1.24.0
seaborn==0.12.0
matplotlib==3.7.0